pyproject.toml
[project]
name = "taskbadger"
version = "1.4.0"
description = "The official Python SDK for Task Badger"
requires-python = ">=3.9"
authors = []
readme = "README.md"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "httpx >=0.20.0,<0.28.0",
    "attrs >=21.3.0",
    "python-dateutil >=2.8.0",
    "typer[all] <0.10.0",
    "tomlkit >=0.12.5",
    "importlib-metadata >=1.0; python_version < '3.8'",
    "typing-extensions >=4.7.1; python_version <= '3.9'",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
include = [
  "taskbadger",
  "taskbadger/internal/py.typed",
]

[project.optional-dependencies]
celery = [
    "celery>=4.0.0,<6.0.0",
]

[tool.uv]
package = true

[project.urls]
"Changelog" = "https://github.com/taskbadger/taskbadger-python/releases"
homepage = "https://taskbadger.net/"
repository = "https://github.com/taskbadger/taskbadger-python"
documentation = "https://docs.taskbadger.net/"


[dependency-groups]
dev = [
    "black",
    "isort",
    "pre-commit",
    "pytest",
    "pytest-httpx",
    "invoke",
    "pytest-celery",
    "redis",
    "openapi-python-client",
]

[project.scripts]
taskbadger = "taskbadger.cli_main:app"

[tool.pytest.ini_options]
# don't run integration tests unless specifically requested
norecursedirs = ".* integration_tests"


[tool.ruff]
exclude = [
    ".venv",
    ".git",
    ".ruff_cache",
]
line-length = 120
indent-width = 4
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "DJ", "PT"]
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.per-file-ignores]
"taskbadger/internal/*" = ["E501"]
